[AAU] Allow starting missions on auth state change
I updated astroAuthInit to take lists of missions that are
passed in to an OnAuthStateChangeMissions object and then added to the
Locator.

This means that in BindAuthState we can now launch & land all the
missions that were added to OnAuthStateChangeMissions at the
appropriate  auth state changed event.

I also removed the redundant UpdateUserState that was being landed in
SignInWithFirebase so that we leave the update to BindAuthState where
the "auth state changed" events are listended to and don't update the
user state twice on sign.