[ATU] Add StubbedFirestoreService, AstroException
I added StubbedFirestoreService so we can easily replace a
service with a stubbed version during tests, so we can avoid flaky
network calls and the like.

I added AstroException so we can setup stubs to throw our own exception,
then expect that specific type of excpetion to have been thrown.