Split astro_error_handling state classes into interface & implementation
We then move the interfaces to astro_types. This is part of a strategy
to use the Dependency Inversion Principle which is very helpful in our
case where we have plugins that share an object whose type is not
known till runtime, but we can have the type implement interfaces and
cast the shared object where we don't have access to it's full type yet.