chore(+): get all tests working (#232)
* Rename WidgetTestHarness.child to innerWidget

Just for clarity and readability

* [AAU] Split state into interfaces/implementations

* [TPR] Update the_process to the new auth structure

* [PDG] Update to new auth structure (interfaces)

* [RSD] Update to new auth structure (interfaces)

* [OML] Update to new auth structure (interfaces)

* Skip tests, commit files generated by pub upgrade

* Split astro_error_handling state classes into interface & implementation

We then move the interfaces to astro_types. This is part of a strategy
to use the Dependency Inversion Principle which is very helpful in our
case where we have plugins that share an object whose type is not
known till runtime, but we can have the type implement interfaces and
cast the shared object where we don't have access to it's full type yet.

* Update our_meals & fix tests

* Fix tests

So we can get to all tests passing locally making it easier to
run all tests locally as a check before submitting a PR

* Move the AppStateAuth interface to astro_types

* Add missionControl parameter to WidgetTestHarness

So we can use a different MissionControl if a test calls for it. I also
moved flutter_test from a dev_dependency to a dependency, as
this package will itself be used as a dev_dependency and now we
can use flutter_test APIs in the lib folder.