Update the_process
I changed the location of where the app was accessing the firestore,
adding "projects/the-process" as we are sharing a firestore for a few
apps now.

I got the "manage organisations" feature going again (by adding a
MenuOption) and fixed some bugs so the whole process of adding &
deleting organisations is working.