chore(+): Move astro init code to relevant plugins
I created a SystemCheck class with members for pre/post land/launch so we
can apply system checks at any point.

I split the PageGenerator into interface & implementation and put the
interface into astro_types so we can use the type without depending
on the full package (eg. for avoiding circular dependencies).

I removed the const constructor from RemoveCurrentRoute in
astro_navigation - we probably can't use const LandingMissions
because we need the type parameter for landingInstructions
(which takes a 'S state' and returns a S) to be set (eg. to AppState).
Dart doesn't allow type parameters to be set when  calling a const
constructor.

I added an initialize function to astro_inspector_screen that sets up
the system check that sends mission reports to the inspector. We add the
system check to the preLaunch & postLand lists so all mission reports
are created and sent at the appropriate point.

I also added an initialize function to astro_error_handling which adds
the page generator for the ErrorReport route to the Locator.

I fixed a bug in DefaultErrorHandlers.handleLaunchError which came from
casting to an interface type. For now we just cast to dynamic and have
tasks to do something better.

I added to the astro_auth initialization function - adding
AuthGatePageState to the AppState, and corresponding page generator to
the Locator, along witht the FirebaseAuthService.

I cleaned up all the AstroBase files in the apps that depend on
astro & plugins - much of it is now redundant becaue the code is
in the plugin initialization functions.