Complete tech_world astro migration & re-architect
I removed the serverController as it was only being used as a way to
publish click locations to the server and now that the networking
service is in the locator we can just access it directly.

We now add the TechWorldGame to the locator and locate it in the build
of the HomeScreen. This was a way to get the app initialization (which
requires a particular order) working but it's probably overuse of the
locator - decoupling the creation and use of the TechWorldGame is
potentially problematic and it might be worth taking the time to look at
alternative patterns.