Add C wrappers & idiomatic Dart that calls via FFI
I built OpenCV as a framework and moved it in to the macos folder. I
also added C wrapper code to flutter_opencv_ffi.cpp and updated
flutter_opencv_ffi.podspec with instructions to use the OpenCV framework
and the C wrapper code as part of the build. This results in the app
binary having access to the C symbols.

I added Dart code in flutter_opencv_ffi.dart that uess FFI to find
the C wrappers - this is Dart wrappers for the C wrappers which means
a dependendee of the plugin can call the C code via nice idiomatic Dart.

I updated the example to use the opencv functions with a simple app that
lets you edit a photo.

I updated the plugin and example to use a MACOSX_DEPLOYMENT_TARGET of
10.13 because there were annoying build warnings and I believe flutter
now requires 10.13 (just hasn't made it to stabke yet).