Add NativeError, extensions, more functionality
I added a NativeError class that keeps a Pointer<git_error> and outputs
the error message in the toString so we can query libgit2 for errors and
throw in Dart.

I added extensions on String and Pointer<Char> to convert between Dart
and C strings.

I cleaned up and added some more functionality to the NativeReference
and NativeRepository classes so we can start working on related app
features.