feat(gvs): NativeRepository & NativeReference (#252)
* Refactor git objects so visuals wrap state objects

* Move relevant code inside a folder named 'native'

The purpose is two-fold:
- puts the libgit2 folder underneath our CMakeLists.txt which I
think makes it easier to combine our glue code and libgit2 into one build
- put the native C code and related code (eg. Dart bindings) in one place

* Move libgit2_bindings for discoverability

* Adjust file references after rearranging folders

* Make a BranchesIterator & replace CommitTreeState

I think it will be simpler to start from the branches and manually walk
each commit tree rather than perform a whole revision walk from
the start.

* Add compiler-opts to ffigen section of pubspec

- '-Wno-nullability-completeness' avoids the large amount of errors from
missing nullability annotations in usr/include/stdlib.h on macos
- '-Inative/libgit2/include' adds native/libgit2/include to the include
search path

* Add a README to the native folder

The README describes what's in the folder, why it's there. It's just a
start but will become a description of how to build the native part of
the app and how to configure the build if desired.

* Remove windows & web targets

As they're not likely to be used soon and we can regenerate them if wanted

* Move libgit2_bindings.dart to native_bindings folder

Just for better discoverability.

* Comment out the refs.h entry-point

I don't understand the consequences of this change yet so leaving
this PR in the base state so we can have a PR with just the change.

* Use NativeFinalizers to manage native memory

I've added a section to the ffigen part of the pubspec. Under
functions -> symbol-address -> include, there is an item for each
'free' function that we use in various git object classes. The 'free'
functions are used in a NativeFinalizers which automatically free native
memory if the Dart object gets GC'd. The extra section in the pubspec
tells ffigen to expose the relevant generated functions.

I created classes to represent 'native' entities: NativeRepository,
NativeReference & BranchIterator. These classes create and mange native
memory for the relevant git object via libgit2.

Finally I updated GitService to create these Dart objects that manage
the underlying native memory, rather than having memory management
directly in the GitService.

* Add NativeError, extensions, more functionality

I added a NativeError class that keeps a Pointer<git_error> and outputs
the error message in the toString so we can query libgit2 for errors and
throw in Dart.

I added extensions on String and Pointer<Char> to convert between Dart
and C strings.

I cleaned up and added some more functionality to the NativeReference
and NativeRepository classes so we can start working on related app
features.