feat(aeh): Add configurable error reports (#255)
I added an ErrorReportSettings member to ErrorReport and
DefaultErrorHandlers now requires an ErrorReportSettings parameter
in handleLandingError & handleLaunchError so that an exception can
tell the handlers how it wants to be displayed.

I created AstroException with members for message and reportSettings so
that anyone can easily create an exception that configures the resulting
error message in different ways.

I added AstroInfoException that configures the error message to just
display the message with no stack trace or mission details. This is
intended to be used for things like incorrect (but expected) user input
where we want to give the user info about what happened but not details
like the stack trace that look like a crash.

The CreateErrorReport landing mission now uses the ErrorReportSettings
to configure the message and we also pass the ErrorReportSettings in to
the DefaultErrorReport so the ErrorReportScreen can decide what to
display (eg. how to display title & stack trace).