chore(ast): Make MissionControl pass reportSettings (#256)
The ErrorHandlers now take an ErrorReportSettings parameter so we check
if the exception is a known type where we can use the exception's report
settings and if not, pass the default. This means the behaviour remains
the same for exiting code but we can have exceptions that configure
the error report.