Add a service for selecting files, add test utils
I added astro_test_utils and wrote a test that uses the WidgetTestHarness
so we can test the individual parts of a feature and mock out causes of
flakiness or parts that require user input.

I moved the code in SelectPackages mission that uses the file selector
plugin into a separate service, put the service in the locator and used
the locator in the mission. This means we can update the locator in a
test to provide a stubbed test double in place of the service, avoiding
the need for user input.