Move parser code into service & test
I moved the code that parses the pubspec into a service (named
ParserService) and replaced said code in the SelectPackages mission
with a call to the service. The point being separation of concerns & to
make the code easier to test.

I added a test that the ParserService correctly turns a pubspec file
into a list of Dependency objects.

I added a design document with a class diagram. I've only added the
Dependency class for now.