feat(pdg): Select package & parse pubspec (#257)
* Move parsing into a mission, paint dependencies

I moved the parsing of the yaml into a AwayMission, which then creates
a LandingMission to update the dependencies state. This is part of
keeping the business logic out of the views and makes everything more
testable, maintainable, etc.

I put an OnStateChangeBuilder at the top of HomeScreen's widget
tree, that rebuilds whenever state.dependencies changes so the
current set of dependencies is always given to the children.

I updated the DependencyGraphPainter to draw circles and paint the
dependency names as a first step toward painting the dependency graph.

* Add a service for selecting files, add test utils

I added astro_test_utils and wrote a test that uses the WidgetTestHarness
so we can test the individual parts of a feature and mock out causes of
flakiness or parts that require user input.

I moved the code in SelectPackages mission that uses the file selector
plugin into a separate service, put the service in the locator and used
the locator in the mission. This means we can update the locator in a
test to provide a stubbed test double in place of the service, avoiding
the need for user input.

* Make FakeFileSelectorService to avoid user input

In tests we can now add FakeFileSelectorService to the Locator in place
of the regular FileSelectorService so we can skip user input.

I also added test-data in the form of a pubspec.yaml file so the tests
can automatically select that file via FakeFileSelectorService.

* Move parser code into service & test

I moved the code that parses the pubspec into a service (named
ParserService) and replaced said code in the SelectPackages mission
with a call to the service. The point being separation of concerns & to
make the code easier to test.

I added a test that the ParserService correctly turns a pubspec file
into a list of Dependency objects.

I added a design document with a class diagram. I've only added the
Dependency class for now.

* Add documentation, refactor to improve structure

I added documentation: class diagrams, wireframes, ubiquitous language
items.

I moved models to the shared folder - still trying to figure out an
appropriate folder structure for this project.

* Add a sequence diagram for parse pubspec sequence

* Add Package model & group parsed data by package

* Refactor: add package_selection section

Trying some reorganising for discoverability.

* Add custom exception for non-package folder opened

* Wrap file-reading in try/catch, throw custom exception

* Replace our custom exception with AstroInfoException

AstroInfoException has the behaviour we want (error report is an info
message without stack trace or details) so we can just set the message
we want which is simpler than making a custom exception.

* Fix tests

* Fix test description