ATY: Rename annotation, refactor AuthState
AuthState now implements AstroState rather than having DefaultAuthState
do the implementing, so AuthState now has copyWith & toJson functions.
The result is that we can use the interface as a type in more places
where we previously couldn't because we wanted to call copyWith & toJson
on the object.

The annotation was renamed for clarity & readability.