AAU: Refactor DefaultAuthState, add factory method
DefaultAuthState now does nothing except implement AstroState, which
now implements AstroState and so has copyWith & toJson functions and
getters for the members.

I added a factory method AstroAuth.initialState() which is a little awkward
but gets around the issue of wanting to use the interface type everywhere
possible (b/c interface can't have constructor). When macros arrive this
shouldn't be difficult to solve... or maybe it's independent of source
gen... pondering...).