the_process, astro_generators/annotations/snippets
I renamed SubclassGenerator to StateImplementationGenerator for
findability.

I updated StateImplementationGenerator to set the class name prefix to
'Generated'  if neither prefix or suffix is supplied by the user.
There is no use for having prefix & suffix both empty as that makes
the interface and implementation names the same, which causes an error.

I changed the annotation name to "GeneratedImplementation" as it's more
readable imo, it annotates the class as having a "generated
implementation".