Simplify API, add basic feature set
In this PR I changed the API surface somehwat - the model and the
tensors are not exposed directly now, and so everything is done
through the interpreter (which holds the model & tensors).

I did some organisoing & renaming for readability/findability, eg.
bindings is now bindingsGlobal in bindings_global.

As part of adding a test to the example app I added a check when
locating the shared library file that loads a local file if environment
indicates running in a test.

To allow the Interpreter to provide info on input & output tensors I
added an extension on Pointer<TfLiteTensor> that creates and returns
a TensorInfo. I then added funcitons in the Interpeter that get the
relevant tensor's native struct and uses the extension to provide a
Dart TensorInfo.

Interpreter now has a set of basic functionality that matches the
example usage given in the C API (found in comments in common.h),
including
  - a delete function which frees all the allocated memory
  - movong raw data into and out of the tensors

I also added an InterpreterOptions class (with just number of threads
for now, but that will be expanded in future).