Add reshaping of input tensor & complete the test
I switched to a different tflite model for the example (”movenet” b/c
I found detailed specs that we can use to check for valid input/out
tensor info & data).

I setup the test to read in a jpeg image file, preprocess it into a
256x256 RGB888 then set up and invoke the interpreter, then print
the output.

I simplified the API surface a bit more, eg. by using reduce access function
prefixes to get/set and made several updates to the API:
- added shape member to TensorInfo that reports the info in the `dims`
member of the tensor C struct
- added `reshapeInputTensor` function to `Interpreter` so we can
adjust the input tensor for models that allow different input shapes
- For now we take `Uint8List` (assumed to be rgb888) input and return
`Float32List` but that will be generalised in a future PR

To improve the error handling a bit I updated TFLiteStatusException
to require an `intro` String that is used to describe what was happening
when the tflite error occurred.