Encapsulate inference & model loading in classes
I made a separat class for extracting the model from the asset bundle,
saving locally and holding on to the file path.

I moved the code for running inference to a separate class. The code is
now in InferenceRunner. For now this is using the Interpreter just as
the CameraView was but in the next step the interpreter will be created
in an Isolate and the runner will send and recieve message with the
Isolate.

I also added an extension on CameraImage that converts the image data to
the form that the input tensor needs. The extension is just for
readability.

We now combine the model loading and inference setup in the MainScreen
then the CameraView asks the InferenceRunner to run inference on each
image that the image stream receives.