refactor(mgc): Encapsulate in classes, use isolate (#304)
* Added the movenet tflite model

* Added Keypoint model to hold output data

* Added logic for interpreting images & viewing results

- a callback passed to `startImageStream` in CameraView that converts
the pixel data from the camera into a format the TFLite plugin can use.
- KeypointViewer, a CustomPaint that paints circles for each keypoint.
- a ValueNotifier to tie the two together.

* Added a profile launch config

* Added workaround for Android 13 bug

The bug will be fixed in an upcoming build

see: https://github.com/flutter/flutter/issues/109769

* Commit automatic changes made by running a build

* Encapsulate inference & model loading in classes

I made a separat class for extracting the model from the asset bundle,
saving locally and holding on to the file path.

I moved the code for running inference to a separate class. The code is
now in InferenceRunner. For now this is using the Interpreter just as
the CameraView was but in the next step the interpreter will be created
in an Isolate and the runner will send and recieve message with the
Isolate.

I also added an extension on CameraImage that converts the image data to
the form that the input tensor needs. The extension is just for
readability.

We now combine the model loading and inference setup in the MainScreen
then the CameraView asks the InferenceRunner to run inference on each
image that the image stream receives.

* Move views into separate folder for findability

* Add isolate based implementation of InferenceRunner

Co-authored-by: Trevor Davis <tzendou@gmail.com>

---------

Co-authored-by: Trevor Davis <tzendou@gmail.com>