fix(mgc): overcapture bug / isolate communication (#305)
Overcapture of class members, where the class creating & using an isolate 
(InferenceRunner) was fixed by making the entry point function a top-level function.

I also added the missing piece of our isolate communication, ie. getting messages sent 
back via the SendPort - we use a StreamQueue and a "busy" flag in order to have the
Isolate only run inference on the incoming frame if there is not already an inference running.