Visualise rgb data sent to interpreter, fix rotation
We capture the rgb data just before it gets pushed into the interpreter
and use a notifier to send the rgb data to a widget that is above the
CameraPreview in a stack, where it can be visualised.

I added a rotation to the KeypointsPaint on Android, it's just a
hardcoded fix for now, as I wasn't able to figure out how to determine
the rotation in code, we'll have to sort that at some point.