Add a symlink to the built dylib in test/
The symlink points to the dylib generated during a debug build so that
the test can open the file from it's local environment (or think it is).

This is a dodgy hack, we will want to at some stage come up with
a more permanent solution. I have added a note to this effect in
the README.