First pass at adding everything for hello, world
My first go at designing an API for the plugin is public interfaces for
each domain object, which will allow for both FFI and JS implementations.

We calloc Def structs as needed, use them to create C memory for
the domain object then free the memory. We only keep an Id for the
domain object and use the box2c globals to act on the objects via the
Id.

Because C enums are weird ffigen doesn't create equivalent Dart enums
so I am manually creating them and using a test to ensure that if the
underlying C enums change we will be alerted.