feat(b2c): Add Dart layer for hello world (#313)
* Rename structs to CamelCase

* Commit generated config files

* Depend on ffi as we now use some methods directly

* Update ffigen - don't rename structs

We now wrap the structs in idiomatic Dart so no need to rename.

* Add a B2Vec2 class that wraps a b2Vec2 struct

* Put everything in one file, use private constructors

* Update box2c, regenerate bindings

* First pass at adding everything for hello, world

My first go at designing an API for the plugin is public interfaces for
each domain object, which will allow for both FFI and JS implementations.

We calloc Def structs as needed, use them to create C memory for
the domain object then free the memory. We only keep an Id for the
domain object and use the box2c globals to act on the objects via the
Id.

Because C enums are weird ffigen doesn't create equivalent Dart enums
so I am manually creating them and using a test to ensure that if the
underlying C enums change we will be alerted.

* Add some docs - relationships b/w domain objects

* Add a test to the example app

* Add geometry.h entrypoint, generate Polygon struct

I'm not sure if this is the right approach but without adding the
entrypoint the Polygon Struct is erased and we get an opaque pointer.

* Rename interfaces.dart to types.dart

I think it's more readable when we import the file as a library in an app.

* Add an enums file with Dart enums

The Dart enums are equivalent to C enums that were erased b/c ffigen
doesn't support C enums at this stage.

I also added a test that the generated class enum is a match to the
manually created Dart enum. The extremely basic class enum that gets
generated is basically useless and real enums give better type safety.

* Add check when loading dylib for test env

When we are in a test, the dylib is opened from the local file system,
in test/

This is not ideal but we are waiting till we move over to static interop
to replace this with the better option of injecting the dylib (b/c it
will be easier when we are using static interop).

* Use part files and a library, add Shape class

In order to have a clean public interface over both FFI and JS Interop,
we have two different implementations and use private constructors to
allow creating the correct type via the shared public interface. The easiest
way I could think of was to put everything in a single library.

I also added methods on World & Body that access the box2c factory
methods.

* Add a symlink to the built dylib in test/

The symlink points to the dylib generated during a debug build so that
the test can open the file from it's local environment (or think it is).

This is a dodgy hack, we will want to at some stage come up with
a more permanent solution. I have added a note to this effect in
the README.

* Update box2c submodule

* chore(b2c): Update box2c submodule & README