chore(noc): Add Notion docs scraper, fix test (#315)
* feat(noc): Add Notion docs scraper

* Fix test

I added a PartialUserObject class which is used when a "user" object
does not have a "type" member, with a comment of where the class fits in.

I added notes to the README on the error code messages we can get from
the response body (and also how we scrape the Notion docs).

I added test data for a page and added the test data to the response in
during the test setup to give a nice smoke test.