Move extracting interaction info to run-dart-code
The interactio-to-pubsub server now just puts the json from the
interaction directly into the pubsub message, then the extraction
with checking for errors while parsing etc. is in the run-dart-code
server. This is more appropriate b/c we can now put the details of
an error result into the Discord response (b/c the Discord response is
created in the run-dart-code server).