feat(dic): Change slash cmd to message cmd (#318)
* Add "create message command" example

* Update - dart:2.19, latest deps

* Remove response from server as pt1 of new response

* fix(dic): read valid parts  of interaction

In discord_interaction_to_pubsub_message we now read only valid pars of
the iinteraction. What is valid has changed now we are using message commands,
ie, there are no longer 'options'.

* Add various exceptions to catch specific errors

* Rename extractInfo to extractMessageCommandInfo

The function now specifically looks for json key-value pairs that are
expected to be present in message command interactions.

I also added a bunch of validation and thorwing of specific exceptions.

* Improve readability of server.dart with type info

* Add ApplicationCommandType enum

* Add some test data and sweet sweet tests

* Clean up run-dart-code server.dart

Add checks for parse errors, extra type info, call
extractMessageCommandInfo on json in pubsub message

* Move extracting interaction info to run-dart-code

The interactio-to-pubsub server now just puts the json from the
interaction directly into the pubsub message, then the extraction
with checking for errors while parsing etc. is in the run-dart-code
server. This is more appropriate b/c we can now put the details of
an error result into the Discord response (b/c the Discord response is
created in the run-dart-code server).

* Check for formatted Dart code in message

I also added a class "InteractionData" that hold the useful interaction
data that was extracted from the pubsub message. Because we will
access the data in multiple spots it's nicer to use and more readable
in a class.

* Fix evaluate function to handle multiple lines